notewolfy: Improvement regarding escape sequences and command behavior (#1)
* notewolfy: Improvement regarding escape sequences and command behavior

[ENHANCEMENTS]:
When the user uses commands with a wrong naming pattern, more helpful messages are displayed, sometimes even nothing was being displayed as feedback.
When using help and you do not specify any command, you will now see a list of commands that exist.

[BUGFIXES]:
The help command was missing the edit command and also help could crash the application when using a non-existent command.
Also, vi was opened when using edit but it should be vim.
When using arrow keys, they could exit the application since escape sequences were not handled properly.

Signed-off-by: RaphSku <rapsku.dev@gmail.com>

* notewolfy: Improvement regarding escape sequences and command behavior

[BUGFIXES]:
Correct in README notewolfy command, missing delimiter. Needed to correct a test since the expected number of dashes is dependent on the maximum length of any displayed
information when using ls ws.

Signed-off-by: RaphSku <rapsku.dev@gmail.com>

* notewolfy: Improvement regarding escape sequences and command behavior

[BUGFIXES]:
Needed to correct a test case for list workspaces because the padding was not correct

Signed-off-by: RaphSku <rapsku.dev@gmail.com>

* notewolfy: Improvement regarding escape sequences and command behavior

[CI]:
Corrected title of steps

Signed-off-by: RaphSku <rapsku.dev@gmail.com>

---------

Signed-off-by: RaphSku <rapsku.dev@gmail.com>